services/backend/app/pyproject.toml
[tool.poetry]
name = "transvib"
version = "0.1.0"
description = ""
authors = ["gunnar.kistner@gmail.com <gunnar.kistner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.137.1"}
tortoise-orm = "^0.19.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
asyncpg = "^0.26.0"
aerich = "^0.7.1"
bcrypt = "^4.0.1"
httpx = "^0.23.0"
pytest = "^7.2.0"
nest-asyncio = "^1.5.6"
gql = "^3.4.0"
requests = "^2.28.1"
aiohttp = "^3.8.3"
pytest-asyncio = "^0.20.1"
networkx = "^2.8.8"
numpy = "^1.23.5"
sympy = "^1.11.1"
plotly = "^5.11.0"
celery = "^5.2.7"
watchdog = "^2.1.9"



[tool.aerich]
tortoise_orm = "app.db.functions.initialise.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
